Make the edge relation of skeletons be defined in terms of adjacency.
This was the contents of the now deleted adjacentE lemma. Moreover, the
definition of adjacency now has the existential quantifier at the top
level for convenience.

When rewriting the broken proofs, an effort was made to use (s)skelP,
which can greatly improve some proofs.

Also, the definition of edges was moved to multigraph.v, together with
the edge_in_set lemma.